refactor: import container classes from library root module
