+++ src/tomodachi_testcontainers/containers/wiremock.py
@@ -9,7 +9,7 @@ from typing import Any
 
 from testcontainers.core.waiting_utils import wait_for_logs
 
-from tomodachi_testcontainers.containers.common import DockerContainer, copy_folder_to_container
+from tomodachi_testcontainers import DockerContainer, copy_folder_to_container
 
 
 class WireMockContainer(DockerContainer):
