+++ tests/containers/common/test_copy_folder_to_container.py
@@ -3,7 +3,7 @@ from typing import Generator, cast
 
 import pytest
 
-from tomodachi_testcontainers.containers import DockerContainer, copy_folder_to_container
+from tomodachi_testcontainers import DockerContainer, copy_folder_to_container
 
 
 class AlpineContainer(DockerContainer):
