+++ tests/services/test_service_sftp.py
@@ -7,8 +7,7 @@ import pytest
 import pytest_asyncio
 from docker.models.images import Image as DockerImage
 
-from tomodachi_testcontainers.containers import TomodachiContainer
-from tomodachi_testcontainers.containers.sftp import SFTPContainer
+from tomodachi_testcontainers import SFTPContainer, TomodachiContainer
 from tomodachi_testcontainers.utils import get_available_port
